// recommendationvalidationtests/typescript-regression-tests/AWSGuruDetectorCommonCDK/lib/model/detector_type.ts
export enum DetectorType {
    Bandit = "Bandit",
    DragonGlass = "DragonGlass",
    Inconsistency = "Inconsistency",
    Infer = "Infer",
    RulesEngine = "RulesEngine",
    RulesEngineCDO = "RulesEngineCDO",
    InputValidation = "InputValidation",
    Partitioning = "Partitioning",
    DragonStone = "DragonStone",
    ByteCodeDecompilation = "ByteCodeDecompilation",
    IRExtraction = "IRExtraction",
    HardCodedSecrets = "HardCodedSecrets",
    Penrose = "Penrose",
    CodeRemediation = "CodeRemediation",
    LLM = "LLM"
}